gui-tor-manager/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
stem>=1.8.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
redis>=5.0.0
motor>=3.3.0
websockets>=12.0
python-multipart>=0.0.6
python-json-logger>=2.0.7
